实现筛选动态分页功能,但存在删除之后不同步页码功能

+++ src/views/article/index.vue
@@ -5,27 +5,31 @@
       <div slot="header" class="clearfix">
         <span>全部图文</span>
       </div>
-      <el-form ref="form" :model="form" label-width="80px">
+      <el-form ref="form" :model="filterParams" label-width="80px">
         <el-form-item label="文章状态">
-          <el-radio-group v-model="form.resource">
-            <el-radio label="全部"></el-radio>
-            <el-radio label="草稿"></el-radio>
-            <el-radio label="待审核"></el-radio>
-            <el-radio label="审核通过"></el-radio>
-            <el-radio label="审核失败"></el-radio>
+          <el-radio-group v-model="filterParams.status">
+            <el-radio label="">全部</el-radio>
+            <!-- 这里之所以用index,是因为接口要求用0-草稿等表示，
+            数组的索引正好满足,这里的indexz将同步到status中-->
+            <el-radio
+              v-for="(item,index) in stateTypes"
+              :key="item.label"
+              :label="index+''"
+            >{{ item.label }}</el-radio>
           </el-radio-group>
         </el-form-item>
         <el-form-item label="频道列表">
-          <el-select v-model="form.region" placeholder="请选择">
-            <el-option label="开发者资讯" value="shanghai"></el-option>
-            <el-option label="ios" value="beijing"></el-option>
-            <el-option label="c++" value="beijing"></el-option>
-            <el-option label="css" value="beijing"></el-option>
+          <!-- 这里接口列表要求传channel_id -->
+          <el-select v-model="filterParams.channel_id" placeholder="请选择">
+            <el-option label="全部" value=""></el-option>
+            <el-option v-for="item in channels" :key="item.id" :label="item.name" :value="item.id"></el-option>
           </el-select>
         </el-form-item>
         <el-form-item label="活动时间">
           <el-date-picker
-            v-model="form.value1"
+            value-format="yyyy-MM-dd"
+            v-model="begin_end_pubdate"
+            @change="handleDateChange"
             type="daterange"
             range-separator="至"
             start-placeholder="开始日期"
@@ -33,7 +37,11 @@
           ></el-date-picker>
         </el-form-item>
         <el-form-item>
-          <el-button type="primary" @click="onSubmit">查询</el-button>
+          <el-button
+            type="primary"
+            @click="onSubmit"
+            :loading="articleLoading"
+          >查询</el-button>
         </el-form-item>
       </el-form>
     </el-card>
@@ -42,7 +50,7 @@
     <!-- 列表 -->
     <el-card class="list-card">
       <div slot="header" class="clearfix">
-        <span>共找到15条符合条件的内容</span>
+        <span>共找到{{ totalCount }}条符合条件的内容</span>
       </div>
       <!-- table表格 -->
       <!-- data用来指点表格的数据 -->
@@ -55,16 +63,25 @@
           <!-- slot-scope是插槽作用域，scope是自己起的名字，scope中有个成员叫row
           scope.row就是当前的遍历项-->
           <template slot-scope="scope">
-            <img width="40" :src="scope.row.cover.images[0]" alt="">
+            <img width="40" :src="scope.row.cover.images[0]" alt>
           </template>
         </el-table-column>
         <el-table-column prop="title" label="标题" width="180"></el-table-column>
         <el-table-column prop="pubdate" label="发布日期" width="180"></el-table-column>
-        <el-table-column prop="status" label="状态"></el-table-column>
+        <el-table-column label="状态">
+          <!-- 组件库搜索tag标签 -->
+          <template slot-scope="scope">
+            <!-- stateTypes[scope.row.status]表示当前行的状态的类型
+            意思是说选择自定义stateTypes数组的某一个对象-->
+            <el-tag
+              :type="stateTypes[scope.row.status].type"
+            >{{ stateTypes[scope.row.status].label }}</el-tag>
+          </template>
+        </el-table-column>
         <el-table-column label="操作">
           <template slot-scope="scope">
-            <el-button type="success">修改</el-button>
-            <el-button type="danger" @click="handleDelete(scope.row)">删除</el-button>
+            <el-button type="success" round>修改</el-button>
+            <el-button type="danger" round @click="handleDelete(scope.row)">删除</el-button>
           </template>
         </el-table-column>
       </el-table>
@@ -73,12 +90,14 @@
       <!-- 数据分页 -->
       <!-- 默认每页10条数据 -->
       <!-- current-change分页组件事件 -->
+      <!-- current-page表示当前页码随着改变而改变，也就是当前页码高亮效果 -->
       <el-pagination
         background
         layout="prev, pager, next"
         :total="totalCount"
         @current-change="handleCurrentPage"
-        :disabled= "articleLoading"
+        :current-page="page"
+        :disabled="articleLoading"
       ></el-pagination>
       <!-- 数据分页 -->
     </el-card>
@@ -107,16 +126,67 @@ export default {
       },
       totalCount: 0,
       articleLoading: false,
-      page: 1
+      page: 1,
+      // 自定义状态数据字段
+      stateTypes: [
+        {
+          type: 'info',
+          label: '草稿'
+        },
+        {
+          type: '',
+          label: '待审核'
+        },
+        {
+          type: 'success',
+          label: '审核通过'
+        },
+        {
+          type: 'warning',
+          label: '审核失败'
+        },
+        {
+          type: 'danger',
+          label: '已删除'
+        }
+      ],
+      // 频道列表
+      channels: [],
+      // 文章查询条件参数，为了方便把参数写在这里
+      filterParams: {
+        // 文章状态
+        status: '',
+        // 频道id
+        channel_id: '',
+        // 开始时间
+        begin_pubdate: '',
+        // 截止时间
+        end_pubdate: ''
+      },
+      // 这个字段没有什么用，只是日期选择器必须v-model绑定一个数据才出发
+      // 所以设置这个字段存储同步的[开始时间，结束时间]
+      begin_end_pubdate: []
     }
   },
 
   created() {
+    // 加载文章列表
     this.loadArticles()
+    // 初始化的时候加载频道列表
+    this.loadChannels()
   },
   methods: {
     loadArticles(page = 1) {
       this.articleLoading = true
+      // 过滤出有效的查询条件数据字段
+      // 空字符串，null,undefined都不会进入
+      const filetrData = {}
+      for (let key in this.filterParams) {
+        if (this.filterParams[key]) {
+          filetrData[key] = this.filterParams[key]
+        }
+      }
+
       this.$http({
         method: 'GET',
         url: '/articles',
@@ -124,10 +194,13 @@ export default {
           // 请求数据的页码，不传默认为1
           page,
           // 请求数据的每页大小，不传默认为10
-          per_page: 10
+          per_page: 10,
+          // 将查询出有效的数据字段拷贝到当前对象
+          ...filetrData
         }
       }).then(data => {
-        // console.log(data)
+        // 查看数据
+        console.log(data)
         // 列表数据
         this.articles = data.results
         // 总记录数
@@ -135,8 +208,22 @@ export default {
         this.articleLoading = false
       })
     },
+    loadChannels() {
+      this.$http({
+        method: 'GET',
+        url: '/channels'
+        // 不需要传参数，默认值带参数
+      }).then(data => {
+        // console.log(data)
+        // 将channels列表赋值给自定义的channels
+        this.channels = data.channels
+      })
+    },
     onSubmit() {
-      console.log('submit了')
+      // console.log('submit了')
+      // 让分页组件的页码回到第一页
+      this.page = 1
+      this.loadArticles()
     },
     handleCurrentPage(page) {
       // 当页码发生改变的时候，请求页码对应的页码
@@ -144,14 +231,40 @@ export default {
       // 当页码改变的时候，请求页码对应的数据
       this.loadArticles(page)
     },
+    // 删除文章列表事件
     handleDelete(article) {
-      this.$http({
-        method: 'DELETE',
-        url: `articles/${article.id}`
-      }).then(data => {
-        // console.log(data)
-        this.loadArticles(this.page)
+      this.$confirm('确认要删除吗？', '删除提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        this.$http({
+          method: 'DELETE',
+          url: `articles/${article.id}`
+        }).then(data => {
+          // console.log(data)
+          // data.total_count--
+          // this.totalCount = data.total_count
+          this.$message({
+            type: 'success',
+            message: '删除成功!'
+          })
+          // 重新加载数据列表
+          this.loadArticles(this.page)
+        })
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        })
       })
+    },
+    // 日期选择组件改变事件
+    // 这里的参数value是什么什么无所谓，element组件库有说明
+    handleDateChange(value) {
+      // console.log(value)
+      this.filterParams.begin_pubdate = value[0]
+      this.filterParams.end_pubdate = value[1]
     }
   }
 }
